🛠️ Add line spacing and Improve builders

+++ koleton-base/src/main/kotlin/koleton/custom/Attributes.kt
@@ -8,27 +8,27 @@ import com.facebook.shimmer.Shimmer
 
 sealed class Attributes {
     abstract val color: Int
-    abstract val cornerRadius: Int
+    abstract val cornerRadius: Float
     abstract val isShimmerEnabled: Boolean
     abstract val shimmer: Shimmer
-    abstract val lineSpacing: Int
+    abstract val lineSpacing: Float
 }
 
 data class RecyclerViewAttributes(
     val view: RecyclerView,
     @ColorInt override val color: Int,
-    @Px override val cornerRadius: Int,
+    @Px override val cornerRadius: Float,
     override val isShimmerEnabled: Boolean,
     override val shimmer: Shimmer,
-    override val lineSpacing: Int,
+    override val lineSpacing: Float,
     @LayoutRes val itemLayout: Int,
     val itemCount: Int
 ): Attributes()
 
 data class SimpleViewAttributes(
     @ColorInt override val color: Int,
-    @Px override val cornerRadius: Int,
+    @Px override val cornerRadius: Float,
     override val isShimmerEnabled: Boolean,
     override val shimmer: Shimmer,
-    override val lineSpacing: Int
+    override val lineSpacing: Float
 ): Attributes()
